Addressed PR comments and moved DAO to CONNECTCoreLib
